Site/Backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
APScheduler==3.10.4
google-generativeai==0.2.2
scikit-learn==1.3.1
numpy==1.24.4
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
nltk==3.8.1
PyMySQL==1.1.0
cryptography==41.0.7
